Add sign out API endpoint
Added a working sign out API endpoint. This required an update to the auth.js file in the routes directory and the auth.js file in the controllers directory. The API simply clears the cookie that was created in the sign in API endpoint and returns a message stating the sign out was successful.