Add API endpoint to require sign in
Added an API endpoint to require a user to sign in before posting. For our app it makes sense to require a user to sign in before making posts to add ingredients to their inventory or add a recipes to their recipe list. This API endpoint ensures that only a user who is signed in can make such a request.