Add authorization and correct error
Added function to the controller\user.js file to verify user can make changes to profile. Also, made a correction from previous push. Previously, the user was required to sign in prior to getting posts, but I meant to require this when creating a post. I swapped the call to the middleware from the router's get function to the router's post function instead. We can add it back again if we end up needing it, but obviously as stated earlier only signed in and authorized users should be able to make changes to the profile.