Add additional functionality to posts
Added some additional API endpoints to  routes/post file. Now a user can update and delete a post in addition to uploading a photo. We can also now search for a post by a user id. I updated the apiDocs file to reflect the changes. Also, made a slight change to the post model. Now we have a updated property which will show when a post was last updated. And we also have a gitignore file which will keep our secrets for us.