Add Forgot/Reset password; Update front end
Added api endpoints to allow users to reset their passwords if they forget them. The user receives an email with a link to the front end to reset the password. Also, users can now upload photos. Keep in mind we can use this feature if we want users to upload a photo of the ingredients and recipes.