Added Ingredient Update & a Check to ensure no double adding ingredients
I added a function for Updating Ingredients in the ingredients.js controllers.

In order to have the update work, I had to add an "ingredientById" function similar to the "userById" function in user.js controllers.

I brought over the packages for formidable

Finally I added a check condition inside the addIngredient function that checks if the ingredient already exists in that users database